% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{affine_fit}
\alias{affine_fit}
\title{affine_fit}
\usage{
affine_fit(
  model = "BS",
  fact_dep = FALSE,
  n_factors = 3,
  data = data_default,
  st_val = 0,
  max_iter = 200,
  tolerance = 0.1,
  wd = 0
)
}
\arguments{
\item{model}{Specific model to be fit. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence (fact_dep=FALSE)}

\item{n_factors}{Number of factors. For some models, these are set by default (e.g. n_factors=3 for AFNS models)}

\item{data}{Table with the average mortality rates (ages on the rows and years on the columns)}

\item{st_val}{Starting value for the parameters. If not set, then default values will be used}

\item{max_iter}{Maximum number of iterations for the Coordinate Ascent estimation algorithm in absence of convergence}

\item{tolerance}{Minimum value of increase of the log-likelihood value before convergence}

\item{wd}{Working directory for saving the partial output of the estimation process}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Estimation of affine mortality models
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
