% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{affine_project}
\alias{affine_project}
\title{affine_project}
\usage{
affine_project(
  model = c("BS", "AFNS", "AFGNS", "AFUNS", "AFRNS", "CIR", "GMk"),
  fact_dep = c(FALSE, TRUE),
  n_factors = 3,
  parameters,
  data,
  years_proj = 1
)
}
\arguments{
\item{model}{Affine model. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (`fact_dep=TRUE`) or independence}

\item{n_factors}{Number of factors. For some models, these are set by default}

\item{parameters}{Starting value for the parameters. If not set, then default values will be used}

\item{data}{Table with the average mortality rates}

\item{years_proj}{Number of years of ahead-projected rates}
}
\value{
A vector of the projected survival rates for each age
}
\description{
Projected survival curves and average mortality rates over n years ahead projections
}
\examples{
BSd_3F_proj <- affine_project(model="BS", fact_dep=TRUE, n_factors=3, parameters=pe_BSd_3F$fit$par_est, data=mu_bar, years_proj = 1)
}
