% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{heatmap_res}
\alias{heatmap_res}
\title{heatmap_res}
\usage{
heatmap_res(residuals, color = TRUE)
}
\arguments{
\item{residuals}{Table of residuals obtainable using std.res}

\item{color}{TRUE if colored (default) or FALSE if black and white}
}
\value{
A heatmap of the residuals
}
\description{
Returns the heatmaps created by using heatmaply
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
