% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{MAPE_age}
\alias{MAPE_age}
\title{MAPE_age}
\usage{
MAPE_age(observed, estimated)
}
\arguments{
\item{observed}{the table of observed average mortality rates}

\item{estimated}{table of the fitted average mortality rates as obtainable through the function mubar_hat}
}
\value{
Returns a vector with the Mean Absolute Percentage Error by age
}
\description{
Calculation of the Mean Absolute Percentage Error by age (i.e. by each row of the average mortality table)
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
