% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{rates2avg}
\alias{rates2avg}
\title{rates2avg}
\usage{
rates2avg(mu)
}
\arguments{
\item{mu}{the table of average mortality rates}
}
\value{
Returns a matrix of the same dimension as mu with average mortality rates
}
\description{
Converts a table with mortality rates into average mortality rates for use in estimation of affine mortality models
}
\examples{
mu_bar_rates <- rates2avg(mu_rates)
}
