% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{residuals_01}
\alias{residuals_01}
\title{residuals_01}
\usage{
residuals_01(observed, estimated)
}
\arguments{
\item{observed}{the table of observed average mortality rates}

\item{estimated}{table of the fitted average mortality rates as obtainable through the function mubar_hat}
}
\value{
Returns a matrix with 0-1 residuals by age and year
}
\description{
Computation of the positive (1) and negative (0) residuals. Useful to understand the presence of the patterns within the residuals
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
