% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{RMSE}
\alias{RMSE}
\title{RMSE}
\usage{
RMSE(observed, estimated)
}
\arguments{
\item{observed}{the table of observed average mortality rates}

\item{estimated}{table of the fitted average mortality rates as obtainable through the function mubar_hat}
}
\value{
Returns a scalar denoting the Root Mean Squared Error from the fitted model
}
\description{
Calculation of the Root Mean Squared Error
}
\examples{
RMSE(mu_bar, fitted_BSd)
}
