% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{std_res}
\alias{std_res}
\title{std_res}
\usage{
std_res(
  model = c("BS", "AFNS", "AFGNS", "AFUNS", "AFRNS", "CIR", "GMk"),
  fact_dep = c(FALSE, TRUE),
  n_factors = 3,
  parameters,
  data
)
}
\arguments{
\item{model}{Affine model. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence}

\item{n_factors}{Number of factors. For some models, these are set by default}

\item{parameters}{Starting value for the parameters. If not set, then default values will be used}

\item{data}{Table with the average mortality rates}
}
\value{
Matrix with the standardized residuals given model and parameters.
}
\description{
Function returning the standardized residuals using the method of Ungolo et. al. (2023)
}
\examples{
std_resid <- std_res(model="BS", fact_dep=TRUE, n_factors=3, parameters=pe_BSd_3F$fit$par_est, data=mu_bar)
}
