% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{xfilter}
\alias{xfilter}
\title{xfilter}
\usage{
xfilter(
  model = "BS",
  fact_dep = FALSE,
  n_factors = 3,
  parameters = 0,
  data = data_default
)
}
\arguments{
\item{model}{Affine model. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence}

\item{n_factors}{Number of factors. For some models, these are set by default}

\item{parameters}{Value of the parameters. If not set, then default values will be used}

\item{data}{Table with the average mortality rates}
}
\value{
A list with the mean of the latent process for the update step (X_t) and for the prediction step (X_t_c) and for the corresponding covariance matrices S_t and S_t_c.
}
\description{
Estimation of mean and variance of the latent states
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
