% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{affine_fit}
\alias{affine_fit}
\title{affine_fit}
\usage{
affine_fit(
  model = c("BS", "AFNS", "AFGNS", "AFUNS", "AFRNS", "CIR", "GMk"),
  fact_dep = c(FALSE, TRUE),
  n_factors = 3,
  data,
  st_val,
  max_iter = 200,
  tolerance = 0.1,
  wd = 0
)
}
\arguments{
\item{model}{Specific model to be fit. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence (fact_dep=FALSE)}

\item{n_factors}{Number of factors. For some models, these are set by default (e.g. n_factors=3 for AFNS models)}

\item{data}{Table with the average mortality rates (ages on the rows and years on the columns)}

\item{st_val}{Starting value for the parameters to be supplied as list}

\item{max_iter}{Maximum number of iterations for the Coordinate Ascent estimation algorithm in absence of convergence. Default value set to 200}

\item{tolerance}{Minimum value of increase of the log-likelihood value before convergence. Default value set to 0.1}

\item{wd}{Working directory for saving the partial output of the estimation process}
}
\value{
A list with components:
\item{model}{ Name of the model}
\item{fit}{ List with parameter estimates, log-likelihood function value, and table of their value at each iteration}
\item{n.parameters}{ Number of parameters of the model}
\item{AIC}{ Value of the Akaike Information Criterion of the model}
\item{BIC}{ Value of the Bayesian Information Criterion of the model}
}
\description{
Estimation of the parameters of affine mortality models using the gradient-free Nelder-Mead algorithm for groups of parameters.
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
