% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{heatmap_res}
\alias{heatmap_res}
\title{heatmap_res}
\usage{
heatmap_res(residuals, color = TRUE)
}
\arguments{
\item{residuals}{Table of residuals obtainable using std.res}

\item{color}{TRUE if colored (default) or FALSE if black and white}
}
\value{
Plot of the residuals heatmap
}
\description{
Returns the heatmaps created by using heatmaply
}
\examples{
std_resid <- std_res(model="AFNS", fact_dep=FALSE, parameters=pe_AFNSi$fit$par_est, data=mu_bar) # - Create standardized residuals
heatmap_res(residuals=std_resid, color=FALSE) # - Get plot
}
