% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{par_cov}
\alias{par_cov}
\title{par_cov}
\usage{
par_cov(
  method = c("MI", "Bootstrap"),
  model = c("BS", "AFNS", "AFGNS", "AFUNS", "AFRNS", "CIR", "GMk"),
  fact_dep = c(FALSE, TRUE),
  n_factors = 3,
  parameters,
  data,
  D_se = 50,
  BS_s = 500,
  t_excl = 4,
  max_iter = 200,
  tolerance = 0.1,
  wd = 0
)
}
\arguments{
\item{method}{MI if Multiple Imputation or Bootstrap if the Bootstrap is desired}

\item{model}{Affine model. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence (fact_dep=FALSE)}

\item{n_factors}{Number of factors. For some models, these are set by default}

\item{parameters}{Starting value for the parameters.}

\item{data}{Table with the average mortality rates}

\item{D_se}{Number of Iterations if Multiple Imputation is used (default set to 50)}

\item{BS_s}{Number of Boostrap samples if this method is used}

\item{t_excl}{Number of time-periods to be excluded for stability of the Bootstrap method}

\item{max_iter}{Maximum number of iterations for the Coordinate Ascent estimation algorithm in absence of convergence}

\item{tolerance}{Minimum value of improvement of the log-likelihood value before convergence}

\item{wd}{Working directory for saving the partial output of the estimation process}
}
\value{
A list with the covariance matrix of the parameter estimates `Cov_par_est` and their standard errors `St_err`
}
\description{
Estimation of the uncertainty of the parameters by Bootstrap or Multiple imputation
}
\examples{
par_unc_MI <- par_cov(method="MI", model="Bootstrap", fact_dep=TRUE, n_factors=3, parameters=pe_BSd_3F$fit$par_est, data=mu_bar, D_se=5, max_iter=10, tolerance=0.1)
}
