% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{prob_neg_mu}
\alias{prob_neg_mu}
\title{prob_neg_mu}
\usage{
prob_neg_mu(
  model = c("BS", "AFNS", "AFGNS", "AFUNS", "AFRNS", "CIR", "GMk"),
  fact_dep = c(FALSE, TRUE),
  n_factors = 3,
  parameters,
  data,
  years_proj = 1,
  n_simulations = 1e+05
)
}
\arguments{
\item{model}{Affine model. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence}

\item{n_factors}{Number of factors. For some models, these are set by default}

\item{parameters}{Starting value for the parameters. If not set, then default values will be used}

\item{data}{Table with the average mortality rates}

\item{years_proj}{Number of years of ahead-projected rates (default 1 year)}

\item{n_simulations}{Number of simulations (default value 100000)}
}
\value{
A vector of probabilities of negative rates for each age considered in the dataset.
}
\description{
Probability of negative rates when projected over `n` years, based on simulated values of the latent variables
}
\examples{
neg_mu_p <- prob_neg_mu(model="BS", fact_dep=TRUE, n_factors=3, parameters=pe_BSd_3F$fit$par_est, data=mu_bar, years_proj = 10, n_simulations=1000)
}
