% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{xfilter}
\alias{xfilter}
\title{xfilter}
\usage{
xfilter(
  model = c("BS", "AFNS", "AFGNS", "AFUNS", "AFRNS", "CIR", "GMk"),
  fact_dep = c(FALSE, TRUE),
  n_factors = 3,
  parameters,
  data
)
}
\arguments{
\item{model}{Affine model. E.g. for the Blackburn-Sherris model we have model="BS", and so on}

\item{fact_dep}{Boolean parameter indicating whether estimate models with factor dependence (fact_dep=TRUE) or independence}

\item{n_factors}{Number of factors. For some models, these are set by default}

\item{parameters}{Value of the parameters. If not set, then default values will be used}

\item{data}{Table with the average mortality rates}
}
\value{
A list with the following components:
\item{X_t}{ Mean of the latent process X(t) for the update step}
\item{X_t_c}{ Mean of the latent process X(t) for the prediction step}
\item{S_t}{ Covariance matrix of the latent process X(t) for the update step}
\item{S_t_c.}{ Covariance matrix of the latent process X(t) for the prediction step}
}
\description{
Estimation of mean and variance of the latent states using the univariate Kalman filtering procedure of Koopman and Durbin (2000)
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
