% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{xsmooth}
\alias{xsmooth}
\title{xsmooth}
\usage{
xsmooth(X_t, X_t_c, S_t, S_t_c, kappa)
}
\arguments{
\item{X_t}{Mean of the filtered latent process X(t) at the time-update step (X_t from the filtering process)}

\item{X_t_c}{Mean of the filtered latent process X(t) at the prediction step (X_t_c from the filtering process)}

\item{S_t}{Covariance of the filtered latent process X(t) at the time-update step (S_t from the filtering process)}

\item{S_t_c}{Covariance of the filtered latent process X(t) at the prediction step (S_t_c from the filtering process)}

\item{kappa}{parameter kappa from the real-world dynamics of the latent variables}
}
\value{
List with mean and covariance of the distribution of thelatent process X(t)
}
\description{
Estimation of mean and covariance of the distribution of the smoothed latent process X(t) using the Rauch-Tung-Striebel smoother
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
