% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffineMortality.R
\name{xsmooth}
\alias{xsmooth}
\title{xsmooth}
\usage{
xsmooth(filterobject, kappa)
}
\arguments{
\item{filterobject}{A `xfilter` object}

\item{kappa}{parameter kappa from the real-world dynamics of the latent variables}
}
\value{
List with mean `X_t_sm` and covariance `S_t_sm` of the distribution of the latent process X(t)
}
\description{
Estimation of mean and covariance of the distribution of the smoothed latent process X(t) using the Rauch-Tung-Striebel procedure
}
\examples{
X_smoothed <- xsmooth(filterobject=X_filtered, kappa=pe_BSd_3F$fit$par_est$kappa)
}
